refactor: Move method to utils
